+++ dist_node/trans/cont.js
@@ -40,7 +40,11 @@ var __o = require("../structure"),
     }))));
     (Instance.lift = (Instance.prototype.lift = (function(t) {
         return new(Instance)((function(k) {
-            return t.chain(k);
+            return t.chain((function(f, g) {
+                return (function(x) {
+                    return f(g(x));
+                });
+            })(trampoline, k));
         }));
     })));
     return Instance;