+++ tests/statet.js
@@ -20,26 +20,21 @@ exports.simple_of = function(test) {
     var c = M.of(3);
     
     test.deepEqual(
-        evalState(c, 's'),
-        [3]);
-    
-    test.deepEqual(
-        execState(c, 's'),
-        ['s']);
+        runState(c, 's'),
+        [{'x': 3, 's': 's'}]);
     
     test.done();
 };
 
-
-exports.lift = function(test) {
+exports.simple_chain = function(test) {
     var c = M.get
         .chain(function(x) {
-            return M.lift(List.of([x, x * 2]));
+            return M.of(x + 1);
         });
     
     test.deepEqual(
-        evalState(c, 3),
-        [3, 6]);
+        runState(c, 3),
+        [{'x': 4, 's': 3}]);
     
     test.done();
 };
@@ -47,10 +42,16 @@ exports.lift = function(test) {
 
 exports.concat = function(test) {
     var c = M.of(3)
-        .concat(M.of(5));
+        .concat(M.of(5))
+        .concat(M.put('x').chain(function(x) { return M.of(10); }))
+        .concat(M.of(4));
     
     test.deepEqual(
-        evalState(c, 's'),
-        [3, 5]);
+        runState(c, 's'),
+        [
+         {x:3, s:'s'},
+         {x:5, s:'s'},
+         {x:10, s:'x'},
+         {x:4, s:'s'}]);
     test.done();
 };